update denoise_particles
Particles which are outside the forceplates are set to np.nan only while they are outside (not for the full duration).
By default, short particles are not removed.
I re-wrote how to remove particles too far from center.
The tracking distance is increased until less than 255 (previously 180) particles are found.